feat(payments): multi-region Stripe + provider-aware management
Phase 1 — Foundation:
- New src/lib/stripe/config.ts keystone with getStripeClient/getPublishableKey/
  getWebhookSecret/getPortalUrl factories keyed by StripeRegion ("us"|"uk")
- Client-safe types extracted to src/lib/stripe/region.ts
- payment_region column added to subscriptions + partnerships (migration
  20260422100000), backfilled to 'us', indexed
- All 6 server-side Stripe instantiation points refactored to getStripeClient(region)
- Webhook split: legacy /api/webhooks/stripe (back-compat) + new /api/webhooks/
  stripe/[region] dynamic route, both delegating to handler.ts
- Region stamped onto every Stripe metadata payload + db row
- New src/lib/paypal/client.ts (cancel, token helpers)

Phase 2 — Provider-aware management:
- /api/stripe/cancel-subscription is now both provider-aware (Stripe vs PayPal)
  and region-aware (correct Stripe account per row)
- Email templates render per-provider/per-region management section via
  renderManagementSection({provider, region}); webhook handler passes options
  to all four customer-facing email types
- Admin /admin/subscriptions table gains Method (STRIPE/PAYPAL) and Region
  (US/UK) columns
- Footer, FAQModal, AboutUsModal render dynamic portal links via new
  src/lib/payments/portals.ts (single source of truth, falls back to legacy
  single Stripe URL when no NEXT_PUBLIC_STRIPE_PORTAL_URL_* is set)

Phase 3 — Checkout region routing:
- getRegionForBeneficiary implements country->region map (UK/GB/IE -> uk
  when configured; default region otherwise)
- /api/stripe derives region from beneficiary location when client doesn't
  pass an explicit region; returns {region, publishableKey} alongside
  clientSecret
- region embedded in success_url/return_url/cancel_url
- CheckoutContent.tsx reads publishable_key from query string and memoizes
  loadStripe() per key; SponsorshipModal forwards key+region from API
  response into checkout URL

Env vars (all optional; legacy unsuffixed STRIPE_* still work as fallback for
the default region):
- STRIPE_DEFAULT_REGION
- STRIPE_SECRET_KEY_{US,UK}, NEXT_PUBLIC_STRIPE_PUBLISHABLE_KEY_{US,UK}
- STRIPE_WEBHOOK_SECRET_{US,UK}, STRIPE_PORTAL_URL_{US,UK}
- NEXT_PUBLIC_STRIPE_PORTAL_URL_{US,UK} (Footer/FAQ/About links)
