+++ src/lib/types/db.types.ts
@@ -351,6 +351,7 @@ export type Database = {
           frequency: Database["public"]["Enums"]["partnership_frequency"]
           id: string
           payment_intent: string | null
+          payment_region: string
           project: Database["public"]["Enums"]["project_type"]
           status: string
           stripe_subscription_id: string | null
@@ -368,6 +369,7 @@ export type Database = {
           frequency: Database["public"]["Enums"]["partnership_frequency"]
           id?: string
           payment_intent?: string | null
+          payment_region?: string
           project?: Database["public"]["Enums"]["project_type"]
           status?: string
           stripe_subscription_id?: string | null
@@ -385,6 +387,7 @@ export type Database = {
           frequency?: Database["public"]["Enums"]["partnership_frequency"]
           id?: string
           payment_intent?: string | null
+          payment_region?: string
           project?: Database["public"]["Enums"]["project_type"]
           status?: string
           stripe_subscription_id?: string | null
@@ -591,6 +594,7 @@ export type Database = {
           email_notification: boolean | null
           id: string
           interval: string | null
+          payment_region: string
           sponsorship_method:
             | Database["public"]["Enums"]["sponsorship_method"]
             | null
@@ -609,6 +613,7 @@ export type Database = {
           email_notification?: boolean | null
           id?: string
           interval?: string | null
+          payment_region?: string
           sponsorship_method?:
             | Database["public"]["Enums"]["sponsorship_method"]
             | null
@@ -627,6 +632,7 @@ export type Database = {
           email_notification?: boolean | null
           id?: string
           interval?: string | null
+          payment_region?: string
           sponsorship_method?:
             | Database["public"]["Enums"]["sponsorship_method"]
             | null
